add XRay tracing using Natchez
this uses the X-Ray backend directly instead of using OTel, because
AWS Lambda supports X-Ray by default, and including an OTel collector is
extra work. It sounds like including the collector can be done by
building it in as a Lambda Layer, but that's not something I've done
before, and I think this will work for now.